% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiralpeds-package.R
\docType{package}
\name{admiralpeds-package}
\alias{admiralpeds}
\alias{admiralpeds-package}
\title{admiralpeds: Pediatrics Extension Package for ADaM in 'R' Asset Library}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolbox for programming Clinical Data Standards Interchange Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, \url{https://www.cdisc.org/standards/foundational/adam}). The package is an extension package of the 'admiral' package for pediatric clinical trials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiralpeds}
  \item \url{https://github.com/pharmaverse/admiralpeds}
}


Other internal: 
\code{\link{find_closest_bin}()},
\code{\link{get_bins}()},
\code{\link{set_bins}()}
}
\author{
\strong{Maintainer}: Fanny Gautier \email{fanny.gautier@cytel.com}

Authors:
\itemize{
  \item Ross Farrugia
  \item Zelos Zhu
  \item Sukalpo Saha
  \item Lina Patil
  \item Samia Kabi
}

Other contributors:
\itemize{
  \item Laura Liao [contributor]
  \item Remigiusz Kudlacz [contributor]
  \item Pierre Wallet [contributor]
  \item Amin Sherzad [contributor]
  \item David Freedman [contributor]
  \item Mahmoud Hamza [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item Cytel [copyright holder, funder]
  \item Pfizer Inc. [copyright holder, funder]
}

}
\concept{internal}
\keyword{internal}
