% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl_peds}
\alias{adsl_peds}
\title{Subject Level Analysis Dataset-pediatrics}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 35 columns.
}
\usage{
adsl_peds
}
\description{
An updated ADaM ADSL dataset with pediatric patients
}
\seealso{
Datasets
\code{\link{dm_peds}},
\code{\link{vs_peds}}
}
\concept{datasets}
\keyword{datasets}
