% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdc_bmiage}
\alias{cdc_bmiage}
\title{CDC BMI-for-age-chart}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 438 rows and 5 columns.
}
\source{
From inst/CDC_growth_charts/bmiage.R
}
\usage{
cdc_bmiage
}
\description{
BMI-for-age charts, 2 to 20.5 years,
L (Box-Cox transformation for normality),
M (Median),
S (coefficient of variation) parameters
by sex and age, age is in months.
}
\seealso{
Other datasets: 
\code{\link{cdc_htage}},
\code{\link{cdc_wtage}}
}
\concept{datasets}
\keyword{datasets}
