% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdc_wtage}
\alias{cdc_wtage}
\title{CDC Weight-for-age-chart}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 436 rows and 5 columns.
}
\source{
From inst/CDC_growth_charts/wtage.R
}
\usage{
cdc_wtage
}
\description{
Weight-for-age charts, 2 to 20 years,
L (Box-Cox transformation for normality),
M (Median),
S (coefficient of variation) parameters
by sex and age, age is in months.
}
\seealso{
Other datasets: 
\code{\link{cdc_bmiage}},
\code{\link{cdc_htage}}
}
\concept{datasets}
\keyword{datasets}
