% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_interp_records.R
\name{derive_interp_records}
\alias{derive_interp_records}
\title{Derive interpolated rows for the CDC charts (>=2 yrs old)}
\usage{
derive_interp_records(dataset, parameter = "WEIGHT")
}
\arguments{
\item{dataset}{Input metadataset

The variables "AGE", "AGEU", "SEX", "L", "M", "S" are expected to be in the dataset
For BMI the additional variables "P95" and "Sigma" are expected to be in the dataset
Note that "AGE" must be in days so that "AGEU" = "days"}

\item{parameter}{CDC/WHO metadata parameter

\emph{Permitted Values}: "WEIGHT", "HEIGHT" or "BMI" only - Must be not NULL

e.g. \code{parameter = "WEIGHT"},
or   \code{parameter = "HEIGHT"},
or   \code{parameter = "BMI"}.}
}
\value{
The input dataset additional interpolated records with the input dataset records
}
\description{
Derive interpolated rows for the CDC charts (>=2 yrs old) by AGE
for the following paramters: HEIGHT, WEIGHT and BMI
}
\examples{
library(admiral)
library(admiralpeds)
library(dplyr)
data(cdc_htage)
cdc_htage \%>\%
mutate(
 SEX = case_when(
    SEX == 1 ~ "M",
    SEX == 2 ~ "F",
    TRUE ~ NA_character_
    ),
  AGE = AGE * 30.4375) \%>\%
# Interpolate the AGE by SEX
group_by(SEX) \%>\%
# Ensure first that Age unit is "DAYS"
do(derive_interp_records(., parameter = "HEIGHT")) \%>\%
ungroup()
}
\seealso{
Metadata
\code{\link{cdc_bmiage}},
\code{\link{cdc_htage}},
\code{\link{cdc_wtage}},
\code{\link{who_bmi_for_age_boys}},
\code{\link{who_bmi_for_age_girls}},
\code{\link{who_hc_for_age_boys}},
\code{\link{who_hc_for_age_girls}},
\code{\link{who_lgth_ht_for_age_boys}},
\code{\link{who_lgth_ht_for_age_girls}},
\code{\link{who_wt_for_age_boys}},
\code{\link{who_wt_for_age_girls}},
\code{\link{who_wt_for_ht_boys}},
\code{\link{who_wt_for_ht_girls}},
\code{\link{who_wt_for_lgth_boys}},
\code{\link{who_wt_for_lgth_girls}}
}
\concept{metadata}
\keyword{metadata}
