% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_params_growth_height.R
\name{derive_params_growth_height}
\alias{derive_params_growth_height}
\title{Derive Anthropometric indicators (Z-Scores/Percentiles-for-Age) based on Standard Growth Charts}
\usage{
derive_params_growth_height(
  dataset,
  sex,
  height,
  height_unit,
  meta_criteria,
  parameter,
  analysis_var,
  set_values_to_sds = NULL,
  set_values_to_pctl = NULL
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified in \code{sex}, \code{age}, \code{age_unit}, \code{parameter}, \code{analysis_var}
are expected to be in the dataset.}

\item{sex}{Sex

A character vector is expected.

Expected Values: \code{M}, \code{F}}

\item{height}{Current Height/length

A numeric vector is expected. Note that this is the actual height at the current visit.}

\item{height_unit}{Lenght/Height Unit
A character vector is expected.

Expected values: 'cm'}

\item{meta_criteria}{Metadata dataset

A metadata dataset with the following expected variables:
\code{HEIGHT}, \code{HEIGHTU}, \code{SEX}, \code{L}, \code{M}, \code{S}

The dataset can be derived from CDC/WHO or user-defined datasets.
The CDC/WHO growth chart metadata datasets are available in the package and will
require small modifications.
\itemize{
\item \code{HEIGHT} - Height/Length
\item \code{HEIGHTU} - Height Unit
\item \code{SEX} - Sex
\item \code{L} - Power in the Box-Cox transformation to normality
\item \code{M} - Median
\item \code{S} - Coefficient of variation
}}

\item{parameter}{Anthropometric measurement parameter to calculate z-score or percentile

A condition is expected with the input dataset \code{VSTESTCD}/\code{PARAMCD}
for which we want growth derivations:

e.g. \code{parameter = VSTESTCD == "WEIGHT"}.

There is CDC/WHO metadata available for Height, Weight, BMI, and Head Circumference available
in the \code{admiralpeds} package.}

\item{analysis_var}{Variable containing anthropometric measurement

A numeric vector is expected, e.g. \code{AVAL}, \code{VSSTRESN}}

\item{set_values_to_sds}{Variables to be set for Z-Scores

The specified variables are set to the specified values for the new
observations. For example,
\verb{set_values_to_sds(exprs(PARAMCD = “BMIASDS”, PARAM = “BMI-for-height z-score”))}
defines the parameter code and parameter.

\emph{Permitted Values}: List of variable-value pairs

If left as default value, \code{NULL}, then parameter not derived in output dataset}

\item{set_values_to_pctl}{Variables to be set for Percentile

The specified variables are set to the specified values for the new
observations. For example,
\verb{set_values_to_pctl(exprs(PARAMCD = “BMIAPCTL”, PARAM = “BMI-for-height percentile”))}
defines the parameter code and parameter.

\emph{Permitted Values}: List of variable-value pair

If left as default value, \code{NULL}, then parameter not derived in output dataset}
}
\value{
The input dataset additional records with the new parameter added.
}
\description{
Derive Anthropometric indicators (Z-Scores/Percentiles-for-Age) based on Standard Growth Charts
for Weight/BMI/Head Circumference by Height/Length
}
\examples{
library(dplyr)
library(lubridate)
library(rlang)
library(admiral)

advs <- dm_peds \%>\%
  select(USUBJID, BRTHDTC, SEX) \%>\%
  right_join(., vs_peds, by = "USUBJID")

# metadata is in cm
who_meta_criteria <- bind_rows(
  (admiralpeds::who_wt_for_ht_boys \%>\%
    mutate(
      SEX = "M",
      height_unit = "cm"
    )
  ),
  (admiralpeds::who_wt_for_ht_girls \%>\%
    mutate(
      SEX = "F",
      height_unit = "cm"
    )
  )
) \%>\%
  rename(HEIGHT = Height)
}
\concept{der_prm_bds_vs}
\keyword{der_prm_bds_vs}
