% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{find_closest_bin}
\alias{find_closest_bin}
\title{Find closest bin}
\usage{
find_closest_bin(param, breaks, labels)
}
\arguments{
\item{param}{Parameter

Vector containing parameter that needs closest bin checking}

\item{breaks}{Breaks created in \code{get_bins()}}

\item{labels}{Labels created in \code{get_bins()}}
}
\value{
detects appropriate bin
}
\description{
Find closest bin
}
