% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_bin_utils.R
\name{get_age_bins}
\alias{get_age_bins}
\title{Get Age Bins of Metadata}
\usage{
get_age_bins(meta_criteria)
}
\arguments{
\item{meta_criteria}{Metadata dataset

A metadata dataset that is the same as that in \code{derive_params_growth_age()}}
}
\value{
list of breaks and labels for age bins
}
\description{
Creates unique age-bins for each record to speed up joining
}
\seealso{
Other internal: 
\code{\link{admiralpeds-package}},
\code{\link{get_bins}()},
\code{\link{set_age_bins}()},
\code{\link{set_bins}()}
}
\concept{internal}
\keyword{internal}
