% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{get_bins}
\alias{get_bins}
\title{Get Bins of Metadata}
\usage{
get_bins(meta_criteria, param)
}
\arguments{
\item{meta_criteria}{Metadata dataset

A metadata dataset that is the same as that in
\code{derive_params_growth_age()}/\code{derive_params_growth_height()}

\describe{
\item{Default value}{none}
}}

\item{param}{Character vector

Accepted values: "AGE"/"HEIGHT_LENGTH"

\describe{
\item{Default value}{none}
}}
}
\value{
list of breaks and labels for age/height bins
}
\description{
Creates unique age/heights-bins for each record to speed up joining
}
\seealso{
Other internal: 
\code{\link{admiralpeds-package}},
\code{\link{find_closest_bin}()},
\code{\link{set_bins}()}
}
\concept{internal}
\keyword{internal}
