% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdc_data.R
\name{get_cdc_data}
\alias{get_cdc_data}
\title{Retrieve CDC growth chart metadataset}
\usage{
get_cdc_data(type)
}
\arguments{
\item{type}{Type of Chart

A character is expected amongst: "height", "weight", "bmi"}
}
\value{
The appropriate growth chart from the CDC data
}
\description{
Retrieves CDC growth chart metadataset
}
\examples{
get_cdc_data(type = "bmi") \%>\% head()
}
\concept{metadata}
\keyword{metadata}
