% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_bin_utils.R
\name{set_age_bins}
\alias{set_age_bins}
\title{Set Age Bins}
\usage{
set_age_bins(age, breaks, labels)
}
\arguments{
\item{age}{Age column of dataset}
}
\value{
Character vector of unique age bins based on metadata
}
\description{
Adds the unique age-bins to each dataframe to have a column for \code{by} in join.
}
\seealso{
Other internal: 
\code{\link{admiralpeds-package}},
\code{\link{get_age_bins}()}
}
\concept{internal}
\keyword{internal}
