% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_utils.R
\name{set_bins}
\alias{set_bins}
\title{Set Age/Height Bins}
\usage{
set_bins(param, breaks, labels)
}
\arguments{
\item{age}{Age column of dataset}
}
\value{
Character vector of unique age/height bins based on metadata
}
\description{
Adds the unique age/height-bins to each dataframe to have a column for \code{by} in join.
}
\seealso{
Other internal: 
\code{\link{admiralpeds-package}},
\code{\link{get_age_bins}()},
\code{\link{get_bins}()},
\code{\link{set_age_bins}()}
}
\concept{internal}
\keyword{internal}
