% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vs_ped}
\alias{vs_ped}
\title{Vital signs Dataset-updated}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 136 rows and 15 columns.
}
\usage{
vs_ped
}
\description{
An updated SDTM VS dataset with anthropometric measurements for pediatric patients
}
\keyword{datasets}
