% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vs_peds}
\alias{vs_peds}
\title{Vital signs Dataset-updated}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 140 rows and 17 columns.
}
\usage{
vs_peds
}
\description{
An updated SDTM VS dataset with anthropometric measurements for pediatric patients
}
\seealso{
Datasets
\code{\link{dm_peds}},
\code{\link{who_bmi_for_age_boys}},
\code{\link{who_bmi_for_age_girls}},
\code{\link{who_hc_for_age_boys}},
\code{\link{who_hc_for_age_girls}},
\code{\link{who_lgth_ht_for_age_boys}},
\code{\link{who_lgth_ht_for_age_girls}},
\code{\link{who_wt_for_age_boys}},
\code{\link{who_wt_for_age_girls}},
\code{\link{who_wt_for_ht_boys}},
\code{\link{who_wt_for_ht_girls}},
\code{\link{who_wt_for_lgth_boys}},
\code{\link{who_wt_for_lgth_girls}}
}
\concept{datasets}
\keyword{datasets}
