% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WHO_weight_for_age_boys}
\alias{WHO_weight_for_age_boys}
\title{WHO weight-for-age for boys}
\format{
A data frame with 1857 rows and 4 variables:
\describe{
\item{\code{Day}}{integer Age in days}
\item{\code{L}}{double Box-Cox transformation for normality}
\item{\code{M}}{double Median}
\item{\code{S}}{double Coefficient of variation}
}
}
\source{
\url{https://www.who.int/tools/child-growth-standards/standards/weight-for-age}
}
\usage{
WHO_weight_for_age_boys
}
\description{
WHO weight-for-age charts for boys from day 0 (birth) to
day 1856
}
\seealso{
Other datasets: 
\code{\link{WHO_bmi_for_age_boys}},
\code{\link{WHO_bmi_for_age_girls}},
\code{\link{WHO_length_height_for_age_boys}},
\code{\link{WHO_length_height_for_age_girls}},
\code{\link{WHO_weight_for_age_girls}},
\code{\link{WHO_weight_for_height_boys}},
\code{\link{WHO_weight_for_height_girls}},
\code{\link{WHO_weight_for_length_boys}},
\code{\link{WHO_weight_for_length_girls}}
}
\concept{datasets}
\keyword{datasets}
