% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WHO_weight_for_height_girls}
\alias{WHO_weight_for_height_girls}
\title{WHO weight-for-height for girls}
\format{
A data frame with 551 rows and 4 variables:
\describe{
\item{\code{Height}}{double Height in cm}
\item{\code{L}}{double Box-Cox transformation for normality}
\item{\code{M}}{double Median}
\item{\code{S}}{double Coefficient of variation}
}
}
\source{
\url{https://www.who.int/tools/child-growth-standards/standards/weight-for-length-height}
}
\usage{
WHO_weight_for_height_girls
}
\description{
WHO weight-for-height charts for girls from 65cm to 120cm
}
\seealso{
Other datasets: 
\code{\link{WHO_bmi_for_age_boys}},
\code{\link{WHO_bmi_for_age_girls}},
\code{\link{WHO_length_height_for_age_boys}},
\code{\link{WHO_length_height_for_age_girls}},
\code{\link{WHO_weight_for_age_boys}},
\code{\link{WHO_weight_for_age_girls}},
\code{\link{WHO_weight_for_height_boys}},
\code{\link{WHO_weight_for_length_boys}},
\code{\link{WHO_weight_for_length_girls}}
}
\concept{datasets}
\keyword{datasets}
