% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{who_wt_for_age_boys}
\alias{who_wt_for_age_boys}
\title{WHO Weight-for-age for boys}
\format{
A data frame with 1857 rows and 4 variables:
\describe{
\item{\code{Day}}{integer Age in days}
\item{\code{L}}{double Box-Cox transformation for normality}
\item{\code{M}}{double Median}
\item{\code{S}}{double Coefficient of variation}
}
}
\source{
\url{https://www.who.int/tools/child-growth-standards/standards/weight-for-age}
}
\usage{
who_wt_for_age_boys
}
\description{
WHO Weight-for-age charts for boys from day 0 (birth) to
day 1856
}
\seealso{
Metadata
\code{\link{cdc_bmiage}},
\code{\link{cdc_htage}},
\code{\link{cdc_wtage}},
\code{\link{derive_interp_records}()},
\code{\link{who_bmi_for_age_boys}},
\code{\link{who_bmi_for_age_girls}},
\code{\link{who_hc_for_age_boys}},
\code{\link{who_hc_for_age_girls}},
\code{\link{who_lgth_ht_for_age_boys}},
\code{\link{who_lgth_ht_for_age_girls}},
\code{\link{who_wt_for_age_girls}},
\code{\link{who_wt_for_ht_boys}},
\code{\link{who_wt_for_ht_girls}},
\code{\link{who_wt_for_lgth_boys}},
\code{\link{who_wt_for_lgth_girls}}
}
\concept{metadata}
\keyword{metadata}
