% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{who_wt_for_ht_boys}
\alias{who_wt_for_ht_boys}
\title{WHO weight-for-height for boys}
\format{
A data frame with 551 rows and 4 variables:
\describe{
\item{\code{Height}}{double Height in cm}
\item{\code{L}}{double Box-Cox transformation for normality}
\item{\code{M}}{double Median}
\item{\code{S}}{double Coefficient of variation}
}
}
\source{
\url{https://www.who.int/tools/child-growth-standards/standards/weight-for-length-height}
}
\usage{
who_wt_for_ht_boys
}
\description{
WHO weight-for-height charts for boys from 65cm to 120cm
}
\seealso{
Datasets
\code{\link{dm_peds}},
\code{\link{vs_peds}},
\code{\link{who_bmi_for_age_boys}},
\code{\link{who_bmi_for_age_girls}},
\code{\link{who_hc_for_age_boys}},
\code{\link{who_hc_for_age_girls}},
\code{\link{who_lgth_ht_for_age_boys}},
\code{\link{who_lgth_ht_for_age_girls}},
\code{\link{who_wt_for_age_boys}},
\code{\link{who_wt_for_age_girls}},
\code{\link{who_wt_for_ht_girls}},
\code{\link{who_wt_for_lgth_boys}},
\code{\link{who_wt_for_lgth_girls}}
}
\concept{datasets}
\keyword{datasets}
