% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{who_wt_for_lgth_boys}
\alias{who_wt_for_lgth_boys}
\title{WHO weight-for-length for boys}
\format{
A data frame with 651 rows and 4 variables:
\describe{
\item{\code{Length}}{double Length in cm}
\item{\code{L}}{double Box-Cox transformation for normality}
\item{\code{M}}{double Median}
\item{\code{S}}{double Coefficient of variation}
}
}
\source{
\url{https://www.who.int/tools/child-growth-standards/standards/weight-for-length-height}
}
\usage{
who_wt_for_lgth_boys
}
\description{
WHO weight-for-length charts for boys from 45cm to 110cm
}
\seealso{
Other datasets: 
\code{\link{who_bmi_for_age_boys}},
\code{\link{who_bmi_for_age_girls}},
\code{\link{who_lgth_ht_for_age_boys}},
\code{\link{who_lgth_ht_for_age_girls}},
\code{\link{who_wt_for_age_boys}},
\code{\link{who_wt_for_age_girls}},
\code{\link{who_wt_for_ht_boys}},
\code{\link{who_wt_for_ht_girls}},
\code{\link{who_wt_for_lgth_girls}}
}
\concept{datasets}
\keyword{datasets}
