% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdc_bmiage}
\alias{cdc_bmiage}
\title{CDC BMI-for-age-chart}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 438 rows and 5 columns.
}
\source{
\url{https://www.cdc.gov/growthcharts/percentile_data_files.htm}
}
\usage{
cdc_bmiage
}
\description{
BMI-for-age charts, 2 to 20.5 years,
\describe{
\item{\code{SEX}}{Sex: 1 = male, 2 = female}
\item{\code{AGE}}{Age in months}
\item{\code{L}}{Box-Cox transformation for normality}
\item{\code{M}}{Median}
\item{\code{S}}{Coefficient of variation}
}
}
\seealso{
Metadata
\code{\link{cdc_htage}},
\code{\link{cdc_wtage}}
}
\concept{metadata}
\keyword{metadata}
