% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdc_htage}
\alias{cdc_htage}
\title{CDC Height-for-age-chart}
\format{
A data frame with 436 rows and 5 variables:
\describe{
\item{\code{SEX}}{Sex: 1 = male, 2 = female}
\item{\code{AGE}}{Age in months}
\item{\code{L}}{Box-Cox transformation for normality}
\item{\code{M}}{Median}
\item{\code{S}}{Coefficient of variation}
}
}
\source{
\url{https://www.cdc.gov/growthcharts/percentile_data_files.htm}
}
\usage{
cdc_htage
}
\description{
Height-for-age charts, 2 to 20 years
}
\seealso{
Metadata
\code{\link{cdc_bmiage}},
\code{\link{cdc_wtage}},
\code{\link{derive_interp_records}()},
\code{\link{who_bmi_for_age_boys}},
\code{\link{who_bmi_for_age_girls}},
\code{\link{who_hc_for_age_boys}},
\code{\link{who_hc_for_age_girls}},
\code{\link{who_lgth_ht_for_age_boys}},
\code{\link{who_lgth_ht_for_age_girls}},
\code{\link{who_wt_for_age_boys}},
\code{\link{who_wt_for_age_girls}},
\code{\link{who_wt_for_lgth_boys}},
\code{\link{who_wt_for_lgth_girls}}
}
\concept{metadata}
\keyword{metadata}
