% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdc_wtage}
\alias{cdc_wtage}
\title{CDC Weight-for-age-chart}
\format{
A data frame with 436 rows and 5 variables:
\describe{
\item{\code{SEX}}{Sex: 1 = male, 2 = female}
\item{\code{AGE}}{Age in months}
\item{\code{L}}{Box-Cox transformation for normality}
\item{\code{M}}{Median}
\item{\code{S}}{Coefficient of variation}
}
}
\source{
\url{https://www.cdc.gov/growthcharts/percentile_data_files.htm}
}
\usage{
cdc_wtage
}
\description{
Weight-for-age charts, 2 to 20 years
}
\seealso{
Metadata
\code{\link{cdc_bmiage}},
\code{\link{cdc_htage}}
}
\concept{metadata}
\keyword{metadata}
