% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdc_wtage}
\alias{cdc_wtage}
\title{CDC weight-for-age}
\format{
A data frame with 436 rows and 5 variables:
\describe{
\item{\code{SEX}}{double sex where 1 = male, 2 = female}
\item{\code{AGE}}{double age in months}
\item{\code{L}}{double Box-Cox transformation for normality}
\item{\code{M}}{double Median}
\item{\code{S}}{double Coefficient of variation}
}
}
\source{
\url{https://www.cdc.gov/growthcharts/percentile_data_files.htm}
}
\usage{
cdc_wtage
}
\description{
CDC weight-for-age charts from age 2 to 20 years (in months)
}
\seealso{
Other datasets: 
\code{\link{cdc_bmiage}},
\code{\link{cdc_htage}}
}
\concept{datasets}
\keyword{datasets}
