% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dm_peds}
\alias{dm_peds}
\title{Demographic Dataset-updated}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 26 columns.
}
\usage{
dm_peds
}
\description{
An updated SDTM DM dataset with pediatric patients
}
\seealso{
Datasets
\code{\link{adsl_peds}},
\code{\link{vs_peds}}
}
\concept{datasets}
\keyword{datasets}
