% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vs_peds}
\alias{vs_peds}
\title{Vital signs Dataset-updated}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 164 rows and 26 columns.
}
\usage{
vs_peds
}
\description{
An updated SDTM VS dataset with anthropometric measurements for pediatric patients
}
\seealso{
Datasets
\code{\link{adsl_peds}},
\code{\link{dm_peds}}
}
\concept{datasets}
\keyword{datasets}
